cli: split `cli_test.go` into separate files
This increases discoverability.

Release note: None
