+++ pkg/cli/BUILD.bazel
@@ -274,6 +274,7 @@ go_test(
     name = "cli_test",
     size = "large",
     srcs = [
+        "cert_test.go",
         "cli_debug_test.go",
         "cli_test.go",
         "connect_join_test.go",
@@ -288,17 +289,23 @@ go_test(
         "doctor_test.go",
         "error_test.go",
         "flags_test.go",
+        "format_table_test.go",
+        "gen_test.go",
         "haproxy_test.go",
         "import_test.go",
         "log_flags_test.go",
         "main_test.go",
+        "node_test.go",
         "nodelocal_test.go",
+        "quit_test.go",
         "sql_format_value_test.go",
         "sql_test.go",
         "sql_util_test.go",
+        "sqlfmt_test.go",
         "start_test.go",
         "statement_diag_test.go",
         "userfiletable_test.go",
+        "workload_test.go",
         "zip_helpers_test.go",
         "zip_test.go",
     ],
@@ -359,6 +366,7 @@ go_test(
         "@com_github_cockroachdb_pebble//:pebble",
         "@com_github_kr_pretty//:pretty",
         "@com_github_lib_pq//:pq",
+        "@com_github_pkg_errors//:errors",
         "@com_github_spf13_cobra//:cobra",
         "@com_github_spf13_pflag//:pflag",
         "@com_github_stretchr_testify//assert",
