+++ pkg/cli/sql_test.go
@@ -16,12 +16,293 @@ import (
 	"os"
 	"testing"
 
+	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/stretchr/testify/assert"
 )
+
+func Example_sql() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	c.RunWithArgs([]string{`sql`, `-e`, `show application_name`})
+	c.RunWithArgs([]string{`sql`, `-e`, `create database t; create table t.f (x int, y int); insert into t.f values (42, 69)`})
+	c.RunWithArgs([]string{`sql`, `-e`, `select 3 as "3"`, `-e`, `select * from t.f`})
+	c.RunWithArgs([]string{`sql`, `-e`, `begin`, `-e`, `select 3 as "3"`, `-e`, `commit`})
+	c.RunWithArgs([]string{`sql`, `-e`, `select * from t.f`})
+	c.RunWithArgs([]string{`sql`, `--execute=SELECT database_name, owner FROM [show databases]`})
+	c.RunWithArgs([]string{`sql`, `-e`, `select 1 as "1"; select 2 as "2"`})
+	c.RunWithArgs([]string{`sql`, `-e`, `select 1 as "1"; select 2 as "@" where false`})
+	// CREATE TABLE AS returns a SELECT tag with a row count, check this.
+	c.RunWithArgs([]string{`sql`, `-e`, `create table t.g1 (x int)`})
+	c.RunWithArgs([]string{`sql`, `-e`, `create table t.g2 as select * from generate_series(1,10)`})
+	// It must be possible to access pre-defined/virtual tables even if the current database
+	// does not exist yet.
+	c.RunWithArgs([]string{`sql`, `-d`, `nonexistent`, `-e`, `select count(*) from "".information_schema.tables limit 0`})
+	// It must be possible to create the current database after the
+	// connection was established.
+	c.RunWithArgs([]string{`sql`, `-d`, `nonexistent`, `-e`, `create database nonexistent; create table foo(x int); select * from foo`})
+	// COPY should return an intelligible error message.
+	c.RunWithArgs([]string{`sql`, `-e`, `copy t.f from stdin`})
+	// --set changes client-side variables before executing commands.
+	c.RunWithArgs([]string{`sql`, `--set=errexit=0`, `-e`, `select nonexistent`, `-e`, `select 123 as "123"`})
+	c.RunWithArgs([]string{`sql`, `--set`, `echo=true`, `-e`, `select 123 as "123"`})
+	c.RunWithArgs([]string{`sql`, `--set`, `unknownoption`, `-e`, `select 123 as "123"`})
+	// Check that partial results + error get reported together. The query will
+	// run via the vectorized execution engine which operates on the batches of
+	// growing capacity starting at 1 (the batch sizes will be 1, 2, 4, ...),
+	// and with the query below the division by zero error will occur after the
+	// first batch consisting of 1 row has been returned to the client.
+	c.RunWithArgs([]string{`sql`, `-e`, `select 1/(@1-2) from generate_series(1,3)`})
+	c.RunWithArgs([]string{`sql`, `-e`, `SELECT '20:01:02+03:04:05'::timetz AS regression_65066`})
+
+	// Output:
+	// sql -e show application_name
+	// application_name
+	// $ cockroach sql
+	// sql -e create database t; create table t.f (x int, y int); insert into t.f values (42, 69)
+	// INSERT 1
+	// sql -e select 3 as "3" -e select * from t.f
+	// 3
+	// 3
+	// x	y
+	// 42	69
+	// sql -e begin -e select 3 as "3" -e commit
+	// BEGIN
+	// 3
+	// 3
+	// COMMIT
+	// sql -e select * from t.f
+	// x	y
+	// 42	69
+	// sql --execute=SELECT database_name, owner FROM [show databases]
+	// database_name	owner
+	// defaultdb	root
+	// postgres	root
+	// system	node
+	// t	root
+	// sql -e select 1 as "1"; select 2 as "2"
+	// 1
+	// 1
+	// 2
+	// 2
+	// sql -e select 1 as "1"; select 2 as "@" where false
+	// 1
+	// 1
+	// @
+	// sql -e create table t.g1 (x int)
+	// CREATE TABLE
+	// sql -e create table t.g2 as select * from generate_series(1,10)
+	// CREATE TABLE AS
+	// sql -d nonexistent -e select count(*) from "".information_schema.tables limit 0
+	// count
+	// sql -d nonexistent -e create database nonexistent; create table foo(x int); select * from foo
+	// x
+	// sql -e copy t.f from stdin
+	// ERROR: woops! COPY has confused this client! Suggestion: use 'psql' for COPY
+	// sql --set=errexit=0 -e select nonexistent -e select 123 as "123"
+	// ERROR: column "nonexistent" does not exist
+	// SQLSTATE: 42703
+	// 123
+	// 123
+	// sql --set echo=true -e select 123 as "123"
+	// > select 123 as "123"
+	// 123
+	// 123
+	// sql --set unknownoption -e select 123 as "123"
+	// invalid syntax: \set unknownoption. Try \? for help.
+	// ERROR: invalid syntax
+	// sql -e select 1/(@1-2) from generate_series(1,3)
+	// ?column?
+	// -1
+	// (error encountered after some results were delivered)
+	// ERROR: division by zero
+	// SQLSTATE: 22012
+	// sql -e SELECT '20:01:02+03:04:05'::timetz AS regression_65066
+	// regression_65066
+	// 20:01:02+03:04:05
+}
+
+func Example_sql_watch() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	c.RunWithArgs([]string{`sql`, `-e`, `create table d(x int); insert into d values(3)`})
+	c.RunWithArgs([]string{`sql`, `--watch`, `.1s`, `-e`, `update d set x=x-1 returning 1/x as dec`})
+
+	// Output:
+	// sql -e create table d(x int); insert into d values(3)
+	// INSERT 1
+	// sql --watch .1s -e update d set x=x-1 returning 1/x as dec
+	// dec
+	// 0.5
+	// dec
+	// 1
+	// ERROR: division by zero
+	// SQLSTATE: 22012
+}
+
+func Example_misc_table() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	c.RunWithArgs([]string{"sql", "-e", "create database t; create table t.t (s string, d string);"})
+	c.RunWithArgs([]string{"sql", "--format=table", "-e", "select '  hai' as x"})
+	c.RunWithArgs([]string{"sql", "--format=table", "-e", "explain select s, 'foo' from t.t"})
+
+	// Output:
+	// sql -e create database t; create table t.t (s string, d string);
+	// CREATE TABLE
+	// sql --format=table -e select '  hai' as x
+	//     x
+	// ---------
+	//     hai
+	// (1 row)
+	// sql --format=table -e explain select s, 'foo' from t.t
+	//            info
+	// --------------------------
+	//   distribution: full
+	//   vectorized: true
+	//
+	//   • render
+	//   │
+	//   └── • scan
+	//         missing stats
+	//         table: t@primary
+	//         spans: FULL SCAN
+	// (9 rows)
+}
+
+func Example_in_memory() {
+	spec, err := base.NewStoreSpec("type=mem,size=1GiB")
+	if err != nil {
+		panic(err)
+	}
+	c := NewCLITest(TestCLIParams{
+		StoreSpecs: []base.StoreSpec{spec},
+	})
+	defer c.Cleanup()
+
+	// Test some sql to ensure that the in memory store is working.
+	c.RunWithArgs([]string{"sql", "-e", "create database t; create table t.f (x int, y int); insert into t.f values (42, 69)"})
+	c.RunWithArgs([]string{"node", "ls"})
+
+	// Output:
+	// sql -e create database t; create table t.f (x int, y int); insert into t.f values (42, 69)
+	// INSERT 1
+	// node ls
+	// id
+	// 1
+	//
+}
+
+func Example_pretty_print_numerical_strings() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	// All strings in pretty-print output should be aligned to left regardless of their contents
+	c.RunWithArgs([]string{"sql", "-e", "create database t; create table t.t (s string, d string);"})
+	c.RunWithArgs([]string{"sql", "-e", "insert into t.t values (e'0', 'positive numerical string')"})
+	c.RunWithArgs([]string{"sql", "-e", "insert into t.t values (e'-1', 'negative numerical string')"})
+	c.RunWithArgs([]string{"sql", "-e", "insert into t.t values (e'1.0', 'decimal numerical string')"})
+	c.RunWithArgs([]string{"sql", "-e", "insert into t.t values (e'aaaaa', 'non-numerical string')"})
+	c.RunWithArgs([]string{"sql", "--format=table", "-e", "select * from t.t"})
+
+	// Output:
+	// sql -e create database t; create table t.t (s string, d string);
+	// CREATE TABLE
+	// sql -e insert into t.t values (e'0', 'positive numerical string')
+	// INSERT 1
+	// sql -e insert into t.t values (e'-1', 'negative numerical string')
+	// INSERT 1
+	// sql -e insert into t.t values (e'1.0', 'decimal numerical string')
+	// INSERT 1
+	// sql -e insert into t.t values (e'aaaaa', 'non-numerical string')
+	// INSERT 1
+	// sql --format=table -e select * from t.t
+	//     s   |             d
+	// --------+----------------------------
+	//   0     | positive numerical string
+	//   -1    | negative numerical string
+	//   1.0   | decimal numerical string
+	//   aaaaa | non-numerical string
+	// (4 rows)
+}
+
+// Example_read_from_file tests the -f parameter.
+// The input file contains a mix of client-side and
+// server-side commands to ensure that both are supported with -f.
+func Example_read_from_file() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	c.RunWithArgs([]string{"sql", "-e", "select 1", "-f", "testdata/inputfile.sql"})
+	c.RunWithArgs([]string{"sql", "-f", "testdata/inputfile.sql"})
+
+	// Output:
+	// sql -e select 1 -f testdata/inputfile.sql
+	// ERROR: unsupported combination: --execute and --file
+	// sql -f testdata/inputfile.sql
+	// SET
+	// CREATE TABLE
+	// > INSERT INTO test(s) VALUES ('hello'), ('world');
+	// INSERT 2
+	// > SELECT * FROM test;
+	// s
+	// hello
+	// world
+	// > SELECT undefined;
+	// ERROR: column "undefined" does not exist
+	// SQLSTATE: 42703
+	// ERROR: column "undefined" does not exist
+	// SQLSTATE: 42703
+}
+
+// Example_includes tests the \i command.
+func Example_includes() {
+	c := NewCLITest(TestCLIParams{})
+	defer c.Cleanup()
+
+	c.RunWithArgs([]string{"sql", "-f", "testdata/i_twolevels1.sql"})
+	c.RunWithArgs([]string{"sql", "-f", "testdata/i_multiline.sql"})
+	c.RunWithArgs([]string{"sql", "-f", "testdata/i_stopmiddle.sql"})
+	c.RunWithArgs([]string{"sql", "-f", "testdata/i_maxrecursion.sql"})
+
+	// Output:
+	// sql -f testdata/i_twolevels1.sql
+	// > SELECT 123;
+	// ?column?
+	// 123
+	// > SELECT 789;
+	// ?column?
+	// 789
+	// ?column?
+	// 456
+	// sql -f testdata/i_multiline.sql
+	// ERROR: at or near "\": syntax error
+	// SQLSTATE: 42601
+	// DETAIL: source SQL:
+	// SELECT -- incomplete statement, \i invalid
+	// \i testdata/i_twolevels2.sql
+	// ^
+	// HINT: try \h SELECT
+	// ERROR: at or near "\": syntax error
+	// SQLSTATE: 42601
+	// DETAIL: source SQL:
+	// SELECT -- incomplete statement, \i invalid
+	// \i testdata/i_twolevels2.sql
+	// ^
+	// HINT: try \h SELECT
+	// sql -f testdata/i_stopmiddle.sql
+	// ?column?
+	// 123
+	// sql -f testdata/i_maxrecursion.sql
+	// \i: too many recursion levels (max 10)
+	// ERROR: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: testdata/i_maxrecursion.sql: \i: too many recursion levels (max 10)
+}
 
 // Example_sql_lex tests the usage of the lexer in the sql subcommand.
 func Example_sql_lex() {
